New FlicketAction structure
Modified FlicketAction structure, not it does not use boolean columns
for storing data, but rather it uses one string column to store
action and one JSON column to store additional data ralated to the
action.

As this si big data change, migration script is prepared to upgrade
database including data transformation and to downgrade database with
data transformation back to original structure.