Bugfix in change category
Change category stopped to work after some modification in the
corresponding view. So adding back the feature change category.

Another question is, if it is correct to only change category and not to
unassign user and set status to open. In my opinion, if someone is
passing ticket to another workgroup, he/she is probably not the part of
that workgroup, so unassign is wellcome, same as setting status to Open,
as the ticket goes to another workteam and the teamleader (or whatever
rules in that team are) is responsible to delegate someone to that
ticket.

But nevermind, at least the category should be changed and this fixes
it.